requirements.txt
langgraph==0.2.0
langchain==0.2.0
langchain-openai==0.1.0
langchain-community==0.2.0
pydantic==2.8.0
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.20
jinja2==3.1.3
python-dotenv==1.0.1
streamlit==1.35.0
pandas==2.2.0
numpy==1.26.0